+++ CHANGELOG.md
@@ -1,8 +1,17 @@
+## Unreleased
+
+- Deprecate 'enable' and configure reporting either statically, via
+  `captureExceptionAction`, or dynamically on per-exception basis via
+  `captureExceptionFilter` (which also allows modifying reported exception).
+- Fallback to current stack trace if the one passed to `captureException` is
+  empty (previously was only for `null` stack traces).
+
 ## 0.4.4
 
 - Reliably cause a native crash on Android.
-- Fix #38 `kotlin.UninitializedPropertyAccessException` in Android plugin
-- Support release health tracking for iOS. This was previously added for Android in 0.4.0
+- Fix #38 `kotlin.UninitializedPropertyAccessException` in Android plugin.
+- Support release health tracking for iOS. This was previously added for Android
+  in 0.4.0.
 
 ## 0.4.3